Break compiler up into smaller pieces
