+++ src/compiler/typescript_to_javascript/cargo_toml_files.ts
@@ -1,4 +1,4 @@
-import { Toml } from '../../bin/types';
+import { Toml } from '../utils/types';
 
 export function generateWorkspaceCargoToml(rootPath: string): Toml {
     return `
