+++ src/compiler/utils/result.ts
